Add content from Zotero via API
There is now another instance of Hugo in the /prebuild folder. This
instance connects to the Zotero API on build.

In /prebuild/layout/index.html the code fetches all sources from Zotero,
and creates new markdown-files if there does not exist a file in
/prebuild/public/sources with the same file name. This is because I will
have to write the content of the new mardownfiles at this location,
unless I manualy move them to ./content/kilder. I have not decided on
this yet. Be careful not to delete the public folder in /prebuild!

These files are connected to the main version through ./config.toml.