Change setup of "kilder" and "læringslogg"
"Kilder" are no longer generated as pages. It is a list, fetched from
Zotero, in the same way as "lenker". All sources are marked with a tag
in Zotero, that is used as an ID in the list. That way I can create
footnotes and target the ID. And the sources uses the ID to look for
posts that reference the source. I now dont have to use two instances of
Hugo.

"Læringslogg" are marked with its own tag synta, to separate them from
other content. That way I can divide the tags.

Added a "nå" or "now" page.